requirements.txt
freezegun==0.3.12
coverage==5.0.1
Django==3.0.2
django-appconf==1.0.3
django-axes==4.5.4
django-cors-headers==3.2.1
django-extensions==2.2.5
django-filter==2.2.0
django-ipware==2.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-autodocs==0.4.4
drf-yasg==1.17.0
factory-boy==2.12.0
gunicorn==20.0.4
PyJWT==1.7.1
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.7.0
python-jwt==3.2.5
swagger-spec-validator==2.4.3
psycopg2-binary
